Move another data class to top-level
* Move GitHubMilestone to top-level within GitHubMilestoneManager.kt
* Simplifies other imports, and we can easily move them separate files if
  we ever want/need to
